Fileslist support for mf demuxer, delemited by ',' and change mf fps to float.
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@5353 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ libmpdemux/mf.h
@@ -5,7 +5,7 @@
 extern int    mf_support;
 extern int    mf_w;
 extern int    mf_h;
-extern int    mf_fps;
+extern float  mf_fps;
 extern char * mf_type;
 
 typedef struct
